Made dfa/nfa creation their own files
